+++ index.d.ts
@@ -30,4 +30,5 @@ export interface TurnstileOptions {
     size?: "normal" | "invisible" | "compact";
     retry?: "auto" | "never";
     "retry-interval"?: number; // up to 15m (900_000) in ms, default is 8s
+    "refresh-expired"?: "auto" | "manual" | "never";
 }