+++ HISTORY.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `refresh-expired`
+
 ## [1.0.1] - 2022-11-25
 
 ### Fixed