+++ HISTORY.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - More appropiate type-hints for container and widget id
 
+### Fixed
+
+- Missing token argument for `expired-callback`
+
 
 ## [1.0.1] - 2022-11-25