+++ HISTORY.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-02-06
+
 ### Added
 
 - `refresh-expired`
@@ -33,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[unreleased]: https://github.com/Le0Developer/turnstile-types/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/Le0Developer/turnstile-types/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/Le0Developer/turnstile-types/releases/tag/v1.1.0
 [1.0.1]: https://github.com/Le0Developer/turnstile-types/releases/tag/v1.0.1
 [1.0.0]: https://github.com/Le0Developer/turnstile-types/releases/tag/v1.0.0